refactor creation helper functions
