+++ ee/clickhouse/views/test/test_clickhouse_funnel_correlation.py
@@ -13,6 +13,7 @@ from freezegun import freeze_time
 from ee.clickhouse.models.event import create_event
 from ee.clickhouse.queries.funnels.funnel_correlation import EventOddsRatioSerialized, FunnelCorrelation
 from ee.clickhouse.test.test_journeys import journeys_for, update_or_create_person
+from posthog.api.test.test_team import create_team
 from posthog.constants import FunnelCorrelationType
 from posthog.models.element import Element
 from posthog.models.person import Person, PersonDistinctId
@@ -625,10 +626,6 @@ class FunnelCorrelationTest(BaseTest):
 @pytest.fixture(autouse=True)
 def clear_django_cache():
     cache.clear()
-
-
-def create_team(organization):
-    return Team.objects.create(name="Test Team", organization=organization)
 
 
 class EventPattern(TypedDict, total=False):
