+++ posthog/api/test/test_event_definition.py
@@ -7,8 +7,10 @@ from django.conf import settings
 from freezegun.api import freeze_time
 from rest_framework import status
 
+from posthog.api.test.test_organization import create_organization
+from posthog.api.test.test_team import create_team
+from posthog.api.test.test_user import create_user
 from posthog.models import Event, EventDefinition, Organization, Team
-from posthog.models.user import User
 from posthog.tasks.calculate_event_property_usage import calculate_event_property_usage_for_team
 from posthog.test.base import APIBaseTest
 
@@ -143,36 +145,6 @@ class TestEventDefinitionAPI(APIBaseTest):
         self.assertEqual(response.json()["count"], 1)
         for item in response.json()["results"]:
             self.assertIn(item["name"], ["watched_movie"])
-
-
-def create_organization(name: str) -> Organization:
-    return Organization.objects.create(name=name)
-
-
-def create_user(email: str, password: str, organization: Organization):
-    return User.objects.create_and_join(organization, email, password)
-
-
-def create_team(organization: Organization) -> Team:
-    """
-    This is a helper that just creates a team. It currently uses the orm, but we
-    could use either the api, or django admin to create, to get better parity
-    with real world  scenarios.
-
-    Previously these tests were running `posthog.demo.create_demo_team` which
-    also does a lot of creating of other demo data. This is quite complicated
-    and has a couple of downsides:
-
-      1. the tests take 30 seconds just to startup
-      2. it makes it difficult to see what data is being used
-    """
-    return Team.objects.create(
-        organization=organization,
-        name="Test team",
-        ingested_event=True,
-        completed_snippet_onboarding=True,
-        is_demo=True,
-    )
 
 
 @dataclasses.dataclass
